+++ src/providers/okta.google.provider.ts
@@ -16,7 +16,13 @@ export class OktaGoogleProvider {
         try {
             logger.info('[OktaGoogleProvider] Registering user', profile);
             let user: IUser;
-            const email: string = profile?.emails[0]?.value || profile?.email;
+            let email: string = profile?.emails[0]?.value || profile?.email;
+
+            // If email field was not provided, use fake email from provider ID and provider
+            if (!email) {
+                email = `${profile.id}@google.com`;
+            }
+
             try {
                 const oktaUser: OktaUser = await OktaService.getOktaUserByEmail(email);
                 user = OktaService.convertOktaUserToIUser(oktaUser);
@@ -30,6 +36,7 @@ export class OktaGoogleProvider {
                     role: 'USER',
                     apps: [],
                     provider: OktaOAuthProvider.GOOGLE,
+                    providerId: profile.id,
                 });
             }